[Improvement-3369][api] Introduce resources, scheduler and taskinstance service interface for clear code (#4766)
* [Improvement-3369][api] Introduce resources, scheduler and taskinstance service interface for clear code
