+++ dolphinscheduler-api/src/main/java/org/apache/dolphinscheduler/api/utils/RegexUtils.java
@@ -44,4 +44,13 @@ public class RegexUtils {
         Matcher isNum = pattern.matcher(str);
         return isNum.matches();
     }
+
+    public static String escapeNRT(String str) {
+        // Logging should not be vulnerable to injection attacks: Replace pattern-breaking characters
+        if (str != null && !str.isEmpty()) {
+            return str.replaceAll("[\n|\r|\t]", "_");
+        }
+        return null;
+    }
+
 }
